Move PositionController logic into ViewController
